+++ ouroboros-network/sim-tests-lib/Test/Ouroboros/Network/LedgerPeers.hs
@@ -8,6 +8,8 @@
 
 module Test.Ouroboros.Network.LedgerPeers where
 
+import Codec.CBOR.FlatTerm
+import Control.Concurrent.Class.MonadSTM.Strict
 import Control.Exception (SomeException (..))
 import Control.Monad.Class.MonadAsync
 import Control.Monad.Class.MonadFork
@@ -17,6 +19,9 @@ import Control.Monad.Class.MonadTime.SI
 import Control.Monad.Class.MonadTimer.SI
 import Control.Monad.IOSim hiding (SimResult)
 import Control.Tracer (Tracer (..), nullTracer, traceWith)
+import Data.Aeson
+import Data.Aeson.Types as Aeson
+import Data.ByteString.Char8 qualified as BS
 import Data.IP qualified as IP
 import Data.List as List (foldl', intercalate, isPrefixOf, nub, sortOn)
 import Data.List.NonEmpty qualified as NonEmpty
@@ -32,9 +37,11 @@ import System.Random
 
 import Network.DNS (Domain)
 
-import Cardano.Slotting.Slot (SlotNo, WithOrigin (..))
-import Control.Concurrent.Class.MonadSTM.Strict
+import Cardano.Binary
+import Cardano.Slotting.Slot (SlotNo (..), WithOrigin (..))
 import Ouroboros.Network.PeerSelection.LedgerPeers
+import Ouroboros.Network.PeerSelection.LedgerPeers.Utils
+           (recomputeRelativeStake)
 import Ouroboros.Network.PeerSelection.RelayAccessPoint
 import Ouroboros.Network.PeerSelection.RootPeersDNS
 import Ouroboros.Network.Testing.Data.Script
@@ -44,12 +51,16 @@ import Test.Tasty
 import Test.Tasty.QuickCheck
 import Text.Printf
 
+
 tests :: TestTree
 tests = testGroup "Ouroboros.Network.LedgerPeers"
   [ testProperty "Pick 100%" prop_pick100
   , testProperty "Pick" prop_pick
-  , testProperty "accBigPoolStake" prop_accBigPoolStake
+  , testProperty "accumulateBigLedgerStake" prop_accumulateBigLedgerStake
+  , testProperty "recomputeRelativeStake" prop_recomputeRelativeStake
   , testProperty "getLedgerPeers invariants" prop_getLedgerPeers
+  , testProperty "LedgerPeerSnapshot CBOR version 1" prop_ledgerPeerSnapshotCBORV1
+  , testProperty "LedgerPeerSnapshot JSON version 1" prop_ledgerPeerSnapshotJSONV1
   ]
 
 newtype ArbitraryPortNumber = ArbitraryPortNumber { getArbitraryPortNumber :: PortNumber }
@@ -97,7 +108,7 @@ newtype ArbitrarySlotNo =
 -- of the tests we run.
 instance Arbitrary ArbitrarySlotNo where
     arbitrary =
-      ArbitrarySlotNo . fromInteger <$> arbitrary
+      ArbitrarySlotNo . SlotNo <$> arbitrary
 
 data StakePool = StakePool {
       spStake :: !Word64
@@ -155,6 +166,93 @@ instance Arbitrary ArbLedgerPeersKind where
     shrink (ArbLedgerPeersKind AllLedgerPeers) = [ArbLedgerPeersKind BigLedgerPeers]
     shrink (ArbLedgerPeersKind BigLedgerPeers) = []
 
+newtype ArbStakeMapOverSource = ArbStakeMapOverSource { getArbStakeMapOverSource :: StakeMapOverSource }
+  deriving Show
+
+instance Arbitrary ArbStakeMapOverSource where
+  arbitrary = do
+    peerSnapshot <-
+      oneof [ pure Nothing, Just <$> genPeerSnapshot ]
+    ledgerWithOrigin <- genWithOrigin
+    ula <- arbitrary
+    ledgerPeers <-
+      case (ula, ledgerWithOrigin) of
+        (Always, _) -> LedgerPeers TooOld . getLedgerPools <$> arbitrary
+        (After slotNo, Origin) | slotNo > 0 -> return BeforeSlot
+        (After afterSlotNo, At atSlotNo)
+          | afterSlotNo <= atSlotNo -> LedgerPeers TooOld . getLedgerPools <$> arbitrary
+        _otherwise -> return BeforeSlot
+    (peerMap, bigPeerMap, cachedSlot) <-
+      return $ case peerSnapshot of
+                 Nothing -> (Map.empty, Map.empty, Nothing)
+                 Just (LedgerPeerSnapshotV1 (At slot, accPools))
+                   -> (Map.fromList accPools, Map.fromList accPools, Just slot)
+                 _otherwise -> error "impossible!"
+    return $ ArbStakeMapOverSource StakeMapOverSource {
+      ledgerWithOrigin,
+      ledgerPeers,
+      peerSnapshot,
+      peerMap,
+      bigPeerMap,
+      ula,
+      cachedSlot }
+    where
+      genWithOrigin = do
+        ArbitrarySlotNo slotNo <- arbitrary
+        return $ if slotNo == 0 then Origin else At slotNo
+      genPeerSnapshot = do
+        slotNo <- At . getPositive <$> arbitrary
+        pools <- accumulateBigLedgerStake . getLedgerPools <$> arbitrary
+        return $ LedgerPeerSnapshotV1 (slotNo, pools)
+
+-- | This test checks whether requesting ledger peers works as intended
+-- when snapshot data is available. For each request, peers must be returned from the right
+-- source - either the ledger or snapshot, depending on whether which source is fresher.
+--
+prop_ledgerPeerSnapshot_requests :: ArbStakeMapOverSource
+                                 -> Property
+prop_ledgerPeerSnapshot_requests ArbStakeMapOverSource {
+      getArbStakeMapOverSource = params@StakeMapOverSource {
+          ledgerWithOrigin,
+          ledgerPeers,
+          peerSnapshot,
+          ula } } =
+  counterexample (unlines
+                   ["Counterexample:", "Ledger slot " ++ show ledgerWithOrigin,
+                    "Ledger pools: " ++ show ledgerPeers,
+                    "Snapshot? :" ++ show peerSnapshot,
+                    "UseLedgerAfter: " ++ show ula]) $
+    let (poolMap, bigPoolMap, _slot) = stakeMapWithSlotOverSource params
+        bigPoolRelays = fmap (snd . snd) . Map.toList $ bigPoolMap
+        poolRelays    = fmap (snd . snd) . Map.toList $ poolMap
+    in case (ledgerWithOrigin, ledgerPeers, peerSnapshot) of
+        (At t, LedgerPeers _ ledgerPools, Just (LedgerPeerSnapshot (At t', snapshotAccStake)))
+          | t' >= t ->
+            snapshotRelays === bigPoolRelays .&&. bigPoolRelays === poolRelays
+          | otherwise ->
+                 bigPoolRelays === ledgerBigPoolRelays
+            .&&.    poolRelays === ledgerRelays
+          where
+            snapshotRelays = fmap (snd . snd) snapshotAccStake
+            ledgerBigPoolRelays   = fmap (snd . snd) (accumulateBigLedgerStake ledgerPools)
+            ledgerRelays = fmap (snd . snd) . Map.toList $ accPoolStake ledgerPools
+
+        (_, LedgerPeers _ ledgerPools, Nothing) ->
+               bigPoolRelays === ledgerBigPoolRelays
+          .&&.    poolRelays === ledgerRelays
+          where
+            ledgerBigPoolRelays = fmap (snd . snd) (accumulateBigLedgerStake ledgerPools)
+            ledgerRelays = fmap (snd . snd) . Map.toList $ accPoolStake ledgerPools
+
+        (_, _, Just (LedgerPeerSnapshot (At t', snapshotAccStake)))
+          | After slot <- ula, t' >= slot ->
+            snapshotRelays === bigPoolRelays .&&. bigPoolRelays === poolRelays
+          where
+            snapshotRelays = fmap (snd . snd) snapshotAccStake
+
+        _otherwise -> bigPoolRelays === [] .&&. poolRelays === []
+
+
 -- | A pool with 100% stake should always be picked.
 prop_pick100 :: Word16
              -> NonNegative Int -- ^ number of pools with 0 stake
@@ -194,7 +292,8 @@ prop_pick100 seed (NonNegative n) (ArbLedgerPeersKind ledgerPeersKind) (MockRoot
                 WithLedgerPeersArgs { wlpRng = rng,
                                       wlpConsensusInterface = interface,
                                       wlpTracer = verboseTracer,
-                                      wlpGetUseLedgerPeers = pure $ UseLedgerPeers Always }
+                                      wlpGetUseLedgerPeers = pure $ UseLedgerPeers Always,
+                                      wlpGetLedgerPeerSnapshot = pure Nothing }
                 (\request _ -> do
                   threadDelay 1900 -- we need to invalidate ledger peer's cache
                   resp <- request (NumberOfPeers 1) ledgerPeersKind
@@ -254,7 +353,8 @@ prop_pick (LedgerPools lps) (ArbLedgerPeersKind ledgerPeersKind) count seed (Moc
                 WithLedgerPeersArgs { wlpRng = rng,
                                       wlpConsensusInterface = interface,
                                       wlpTracer = verboseTracer,
-                                      wlpGetUseLedgerPeers = pure $ UseLedgerPeers (After 0) }
+                                      wlpGetUseLedgerPeers = pure $ UseLedgerPeers (After 0),
+                                      wlpGetLedgerPeerSnapshot = pure Nothing }
                 (\request _ -> do
                   threadDelay 1900 -- we need to invalidate ledger peer's cache
                   resp <- request (NumberOfPeers count) ledgerPeersKind
@@ -307,10 +407,9 @@ prop_pick (LedgerPools lps) (ArbLedgerPeersKind ledgerPeersKind) count seed (Moc
                                         === fromIntegral count `min` numOfPeers)
 
 
-prop_accBigPoolStake :: LedgerPools -> Property
-prop_accBigPoolStake  (LedgerPools [])        = property True
-prop_accBigPoolStake  (LedgerPools lps@(_:_)) =
-
+prop_accumulateBigLedgerStake :: LedgerPools -> Property
+prop_accumulateBigLedgerStake  (LedgerPools [])        = property True
+prop_accumulateBigLedgerStake  (LedgerPools lps@(_:_)) =
          -- the accumulated map is non empty, whenever ledger peers set is non
          -- empty
          not (Map.null accumulatedStakeMap)
@@ -324,7 +423,7 @@ prop_accBigPoolStake  (LedgerPools lps@(_:_)) =
              >= unAccPoolStake bigLedgerPeerQuota)
 
          -- This property checks that elements of
-         -- `accBigPoolStake` form an initial sub-list of the ordered ledger
+         -- `accBigPoolStakeMap` form an initial sub-list of the ordered ledger
          -- peers by stake (from large to small).
          --
          -- We relay on the fact that `Map.elems` returns a list of elements
@@ -335,6 +434,37 @@ prop_accBigPoolStake  (LedgerPools lps@(_:_)) =
           $ elems `isPrefixOf` lps'
   where
     accumulatedStakeMap = accBigPoolStakeMap lps
+
+-- |This functions checks the following properties:
+-- 1. The accumulated relative stake adds up to unity
+-- 2. No pool relative stake can be less than 0
+-- 3. The relays aren't mangled
+-- 4. Running this function multiple times always produces the same result
+--
+prop_recomputeRelativeStake :: LedgerPools -> Property
+prop_recomputeRelativeStake (LedgerPools []) = property True
+prop_recomputeRelativeStake (LedgerPools lps) = property $ do
+  lpk <- genLedgerPeersKind
+  let (accStake, relayAccessPointsUnchangedNonNegativeStake) = go (reStake lpk) lps (0, True)
+  return $     counterexample "recomputeRelativeStake: relays modified or negative pool stake calculated"
+                             relayAccessPointsUnchangedNonNegativeStake
+          .&&. accStake === 1
+          .&&. counterexample "violates idempotency"
+                              ((recomputeRelativeStake BigLedgerPeers . recomputeRelativeStake BigLedgerPeers $ lps) == recomputeRelativeStake BigLedgerPeers lps)
+  where
+    genLedgerPeersKind = elements [AllLedgerPeers, BigLedgerPeers]
+    reStake lpk = recomputeRelativeStake lpk lps
+    -- compare relay access points in both lists for equality
+    -- where we assume that recomputerelativestake doesn't change
+    -- the order, and sum up relative stake to make sure it adds up to 1
+    go ((normPoolStake, raps):rest) ((_, raps'):rest') (accStake, _) =
+      if raps == raps' && normPoolStake >= 0
+      then go rest rest' (accStake + normPoolStake, True)
+      else (accStake + normPoolStake, False)
+    go [] (_:_) (accStake, _) = (accStake, False)
+    go (_:_) [] (accStake, _) = (accStake, False)
+    go _ _ (accStake, relayAccessPointsUnchangedNonNegativeStake) = (accStake, relayAccessPointsUnchangedNonNegativeStake)
+
 
 prop_getLedgerPeers :: ArbitrarySlotNo
                     -> ArbitraryLedgerStateJudgement
@@ -368,6 +498,88 @@ prop_getLedgerPeers (ArbitrarySlotNo curSlot)
                   (pure $ curSlotWO)
                   (pure lsj)
                   (pure (Map.elems (accPoolStake lps)))
+
+-- | Checks validity of LedgerPeerSnapshot CBOR encoding, and whether
+--   round trip cycle is the identity function
+--
+prop_ledgerPeerSnapshotCBORV1 :: ArbitrarySlotNo
+                              -> LedgerPools
+                              -> Property
+prop_ledgerPeerSnapshotCBORV1 slotNo
+                              ledgerPools =
+  counterexample (show snapshot) $
+         counterexample ("Invalid CBOR encoding" <> show encoded)
+                        (validFlatTerm encoded)
+    .&&. either ((`counterexample` False) . ("CBOR decode failed: " <>))
+                (counterexample . ("CBOR round trip failed: " <>) . show <*> (snapshot ==))
+                decoded
+  where
+    snapshot = snapshotV1 slotNo ledgerPools
+    encoded = toFlatTerm . toCBOR $ snapshot
+    decoded = fromFlatTerm fromCBOR encoded
+
+-- | Tests if LedgerPeerSnapshot JSON round trip is the identity function
+--
+prop_ledgerPeerSnapshotJSONV1 :: ArbitrarySlotNo
+                              -> LedgerPools
+                              -> Property
+prop_ledgerPeerSnapshotJSONV1 slotNo
+                              ledgerPools =
+  counterexample (show snapshot) $
+     either ((`counterexample` False) . ("JSON decode failed: " <>))
+            (counterexample . ("JSON round trip failed: " <>) . show <*> nearlyEqualModuloFullyQualified snapshot)
+            roundTrip
+  where
+    snapshot = snapshotV1 slotNo ledgerPools
+    roundTrip = case fromJSON . toJSON $ snapshot of
+                  Aeson.Success s -> Right s
+                  Error str       -> Left str
+
+    nearlyEqualModuloFullyQualified snapshotOriginal snapshotRoundTripped =
+      let LedgerPeerSnapshotV1 (wOrigin, relaysWithAccStake) = snapshotOriginal
+          strippedRelaysWithAccStake = stripFQN <$> relaysWithAccStake
+          LedgerPeerSnapshotV1 (wOrigin', relaysWithAccStake') = snapshotRoundTripped
+          strippedRelaysWithAccStake' = stripFQN <$> relaysWithAccStake'
+      in
+             wOrigin === wOrigin'
+        .&&. counterexample "fully qualified name"
+                            (strippedRelaysWithAccStake === strippedRelaysWithAccStake')
+        .&&. counterexample "approximation error"
+                            (compareApprox relaysWithAccStake relaysWithAccStake')
+
+    stripFQN (_, (_, relays)) = step <$> relays
+    step it@(RelayAccessDomain domain port) =
+      case BS.unsnoc domain of
+        Just (prefix, '.') -> RelayAccessDomain prefix port
+        _otherwise         -> it
+    step it = it
+
+    compareApprox left right =
+      let left' = [(accStake, relativeStake)
+                  | (AccPoolStake accStake, (PoolStake relativeStake, _)) <- left]
+          right' = [(accStake, relativeStake)
+                   | (AccPoolStake accStake, (PoolStake relativeStake, _)) <- right]
+          go (accStake, relativeStake)
+             (accStake', relativeStake') =
+               accStake' / accStake > 999999 % 1000000
+            && accStake' / accStake < 1000001 % 1000000
+            && relativeStake' / relativeStake > 999999 % 1000000
+            && relativeStake' / relativeStake < 1000001 % 1000000
+
+      in all (uncurry go) (zip left' right')
+
+-- | helper functions for ledgerpeersnapshot encoding tests
+--
+snapshotV1 :: ArbitrarySlotNo
+           -> LedgerPools
+           -> LedgerPeerSnapshot
+snapshotV1 (ArbitrarySlotNo slot)
+           (LedgerPools pools) = LedgerPeerSnapshotV1 (originOrSlot, poolStakeWithAccumulation)
+  where
+    poolStakeWithAccumulation = Map.assocs . accPoolStake $ pools
+    originOrSlot = if slot == 0
+                   then Origin
+                   else At slot
 
 -- TODO: Belongs in iosim.
 data SimResult a = SimReturn a [String]