+++ ouroboros-network/src/Ouroboros/Network/Diffusion/P2P.hs
@@ -113,8 +113,8 @@ import Ouroboros.Network.PeerSelection.Governor.Types
 #endif
 import Ouroboros.Network.PeerSelection.LedgerPeers (TraceLedgerPeers,
            WithLedgerPeersArgs (..))
-import Ouroboros.Network.PeerSelection.LedgerPeers.Type
-           (LedgerPeersConsensusInterface (..), UseLedgerPeers)
+import Ouroboros.Network.PeerSelection.LedgerPeers.Type (LedgerPeerSnapshot,
+           LedgerPeersConsensusInterface (..), UseLedgerPeers)
 #ifdef POSIX
 import Ouroboros.Network.PeerSelection.PeerMetric (PeerMetrics,
            fetchynessBlocks, upstreamyness)
@@ -245,17 +245,23 @@ nullTracers =
 data ArgumentsExtra m = ArgumentsExtra {
       -- | selection targets for the peer governor
       --
-      daPeerSelectionTargets  :: PeerSelectionTargets
+      daPeerSelectionTargets   :: PeerSelectionTargets
 
-    , daReadLocalRootPeers    :: STM m (LocalRootPeers.Config RelayAccessPoint)
-    , daReadPublicRootPeers   :: STM m (Map RelayAccessPoint PeerAdvertise)
-    , daReadUseBootstrapPeers :: STM m UseBootstrapPeers
+    , daReadLocalRootPeers     :: STM m (LocalRootPeers.Config RelayAccessPoint)
+    , daReadPublicRootPeers    :: STM m (Map RelayAccessPoint PeerAdvertise)
+    , daReadUseBootstrapPeers  :: STM m UseBootstrapPeers
+    -- | Depending on configuration, node may provide us with
+    -- a snapshot of big ledger peers taken at some slot on the chain.
+    -- These peers may be selected by ledgerPeersThread when requested
+    -- by the peer selection governor when the node is syncing up.
+    -- This is especially useful for Genesis consensus mode.
+    , daReadLedgerPeerSnapshot :: STM m (Maybe LedgerPeerSnapshot)
 
     -- | Peer's own PeerSharing value.
     --
     -- This value comes from the node's configuration file and is static.
-    , daOwnPeerSharing        :: PeerSharing
-    , daReadUseLedgerPeers    :: STM m UseLedgerPeers
+    , daOwnPeerSharing         :: PeerSharing
+    , daReadUseLedgerPeers     :: STM m UseLedgerPeers
 
       -- | Timeout which starts once all responder protocols are idle. If the
       -- responders stay idle for duration of the timeout, the connection will
@@ -266,7 +272,7 @@ data ArgumentsExtra m = ArgumentsExtra {
       --
       -- See 'serverProtocolIdleTimeout'.
       --
-    , daProtocolIdleTimeout   :: DiffTime
+    , daProtocolIdleTimeout    :: DiffTime
 
       -- | Time for which /node-to-node/ connections are kept in
       -- 'TerminatingState', it should correspond to the OS configured @TCP@
@@ -276,21 +282,21 @@ data ArgumentsExtra m = ArgumentsExtra {
       -- purpose is to be resilient for delayed packets in the same way @TCP@
       -- is using @TIME_WAIT@.
       --
-    , daTimeWaitTimeout       :: DiffTime
+    , daTimeWaitTimeout        :: DiffTime
 
       -- | Churn interval between churn events in deadline mode.  A small fuzz
       -- is added (max 10 minutes) so that not all nodes churn at the same time.
       --
       -- By default it is set to 3300 seconds.
       --
-    , daDeadlineChurnInterval :: DiffTime
+    , daDeadlineChurnInterval  :: DiffTime
 
       -- | Churn interval between churn events in bulk sync mode.  A small fuzz
       -- is added (max 1 minute) so that not all nodes churn at the same time.
       --
       -- By default it is set to 300 seconds.
       --
-    , daBulkChurnInterval     :: DiffTime
+    , daBulkChurnInterval      :: DiffTime
     }
 
 --
@@ -637,6 +643,7 @@ runM Interfaces
        , daTimeWaitTimeout
        , daDeadlineChurnInterval
        , daBulkChurnInterval
+       , daReadLedgerPeerSnapshot
        }
      Applications
        { daApplicationInitiatorMode
@@ -986,7 +993,8 @@ runM Interfaces
                                          wlpRng = ledgerPeersRng,
                                          wlpConsensusInterface = daLedgerPeersCtx,
                                          wlpTracer = dtTraceLedgerPeersTracer,
-                                         wlpGetUseLedgerPeers = daReadUseLedgerPeers }
+                                         wlpGetUseLedgerPeers = daReadUseLedgerPeers,
+                                         wlpGetLedgerPeerSnapshot = daReadLedgerPeerSnapshot }
 
           peerSelectionGovernor'
             :: forall (muxMode :: MuxMode) b.